requirements.txt
google-api-python-client
httplib2
oauth2client